pyproject.toml
[project]
name = "etf-trend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "JeremyGu2021", email = "jgu@apra.com.au" }
]
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "matplotlib>=3.10.8",
    "mplfinance>=0.12.10b0",
    "numpy>=2.2.6",
    "openai>=2.14.0",
    "pandas>=2.3.3",
    "pyarrow>=22.0.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.7.2",
    "scipy>=1.15.3",
    "statsmodels>=0.14.6",
    "uvicorn>=0.40.0",
    "yfinance>=1.0",
]

[project.scripts]
etf-trend = "etf_trend:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "black>=25.12.0",
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
    "pytest>=9.0.2",
]


[tool.black]
line-length = 100
target-version = ["py310"]
